+++ react-app/src/components/Plot/HelloPlot.js
@@ -3,44 +3,10 @@ import {useState, useEffect} from 'react';
 import Plot from 'react-plotly.js';
 import { useSelector } from 'react-redux';
 import { Redirect } from "react-router";
-
-class Simulation {
-  constructor(stop,interval,fn=x=>x) {
-    this.start = 0;
-    this.stop = stop;
-    this.interval = interval;
-    this.fn = fn;
-    this.domain = [...Array(interval).keys()];
-    this.range = [];
-  }
-
-  map_domain(domain) {
-    return domain.map(x => this.fn(x))
-  }
-
-  proceed() {
-    if(this.start === 0) this.range = this.map_domain(this.domain);
-
-    const live = this.range[this.range.length-1];
-
-    const domain = [];
-
-    for(let x = this.start; x < this.interval; x++)
-      domain.push(x);
-
-    this.domain = domain;
-    this.range = [...this.range.slice(1), this.fn(live)];
-
-    this.start++;
-    this.interval++;
-
-    return {domain: this.domain, range: this.range}
-  }
-}
+import { Simulation } from "../../utilities/simulation.js";
 
 const HelloPlot = () => {
   const user = useSelector(state => state.session.user)
-  const delay = (time) => new Promise(resolve => setTimeout(resolve,time));
 
   const rand_walk = x => {
     const sgn = Math.pow(-1, Math.floor(2*Math.random()));
